provider/azurerm: Add `azurerm_storage_account`
This is an unusual resource (so far) in that it cannot be created in one
call, and instead must be created and the modified to set some of the
parameters.

It appears we may have to play around with timeouts here as well -
perhaps we should allow the maximum timeouts to be set as part of the
provider configuration early on, with very high defaults.